requirements.txt
certifi==2018.8.13
emoji==0.5.0
future==0.16.0
PySocks==1.6.8
python-telegram-bot==10.1.0
clarifai==2.2.3
dnspython==1.15.0
pymongo==3.7.1